Iteration #3 - Desktop Pre-Release

+++ View/SubControls/ProdutosView.xaml.cs
@@ -1,7 +1,10 @@
 ﻿using ModelLib;
+using ModelLib.Adapters;
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows;
@@ -27,14 +30,14 @@ namespace View.UserControls {
 
         public ProdutosView() {
             InitializeComponent();
-
-            if (Controller.LoggedUser.GetType() == typeof(Administrador)
-                || Controller.LoggedUser.Permissao.Contains(Controller.Permissoes["GerenciarProdutos"])) {
+            if (Controller.LoggedUser is Administrador
+                || Controller.LoggedUser.Permissao.FirstOrDefault(o => o.Descricao == Controller.Permissoes["GerenciarProdutos"].Descricao) != null) {
                 ButtonAdicionar.IsEnabled = true;
                 ButtonDeletar.IsEnabled = true;
+                ButtonEditar.IsEnabled = true;
             }
-            if (Controller.LoggedUser.GetType() == typeof(Administrador)
-                || Controller.LoggedUser.Permissao.Contains(Controller.Permissoes["GerenciarEstoque"])) {
+            if (Controller.LoggedUser is Administrador
+                || Controller.LoggedUser.Permissao.FirstOrDefault(o => o.Descricao == Controller.Permissoes["GerenciarEstoque"].Descricao) != null) {
                 ButtonEstoque.IsEnabled = true;
             }
         }
@@ -51,14 +54,31 @@ namespace View.UserControls {
 
         private void Update(object sender, RoutedEventArgs e) {
             var context = new ERPDBModelContainer();
-            DataGridProduto.ItemsSource = (context.ProdutoSet.ToList().Count > 0) ? context.ProdutoSet.ToList() : null;
+            var l = new List<ProdutoAdapter>();
+            foreach (var p in context.ProdutoSet.ToList()) {
+                l.Add(new ProdutoAdapter(p));
+            }
+            DataGridProduto.ItemsSource = (l.Count > 0) ? l : null;
         }
 
         private void ButtonDeletar_Click(object sender, System.Windows.RoutedEventArgs e) {
             var context = new ERPDBModelContainer();
-            foreach (Produto prod in DataGridProduto.SelectedItems) {
-                var p = context.ProdutoSet.Single(o => o.Codigo == prod.Codigo);
-                context.ProdutoSet.Remove(p);
+            foreach (ProdutoAdapter prod in DataGridProduto.SelectedItems) {
+                                        var p = context.ProdutoSet.Single(o => o.Codigo == prod.Codigo);
+                if (p.ProdutoVendaItem.Count > 0) {
+                    if (MessageBox.Show("O produto " + p.Nome
+                        + " está presente em uma ou mais vendas, se remover este produto as vendas não "
+                        + "irão apresentar mais o produto. Deseja continuar?", "Autorização",
+                        MessageBoxButton.YesNo, MessageBoxImage.Question) == MessageBoxResult.Yes) {
+
+                        p.ProdutoVendaItem.Clear();
+                        context.ProdutoSet.Remove(p);
+
+
+                    }
+                } else {
+                    context.ProdutoSet.Remove(p);
+                }
             }
             context.SaveChanges();
             Update(sender, e);
@@ -77,20 +97,29 @@ namespace View.UserControls {
         }
 
         private void UserControl_Loaded(object sender, RoutedEventArgs e) {
-            var ctx = new ERPDBModelContainer();
+            //var ctx = new ERPDBModelContainer();
             Update(sender, e);
-            ComboBoxExibir.ItemsSource = (ctx.CategoriaSet.ToList().Count > 0) ? ctx.CategoriaSet.ToList() : null;
         }
 
         private void DataGridProduto_CellEditEnding(object sender, DataGridCellEditEndingEventArgs e) {
             if (e.EditAction == DataGridEditAction.Commit) {
                 var context = new ERPDBModelContainer();
-                var produto = (Produto)e.Row.Item;
-                var original = context.ProdutoSet.Find(produto.Id);
+                //var original = (Produto)e.Row.Item;
+                var original = context.ProdutoSet.Find(((Produto)e.Row.Item).Id);
+                //var novo = original;
+                ///try {
+                foreach (PropertyInfo prop in typeof(Produto).GetProperties()) {
+                    if (prop.Name == e.Column.Header.ToString()) {
+                        if (prop.PropertyType == typeof(String)) prop.SetValue(original, ((TextBox)e.EditingElement).Text);
+                        else if (prop.PropertyType == typeof(Double)) prop.SetValue(original, Double.Parse(((TextBox)e.EditingElement).Text, NumberStyles.Currency, CultureInfo.InvariantCulture));
+                    }
+                }
+
                 if (original != null) {
-                    context.Entry(original).CurrentValues.SetValues(produto);
+                    //context.Entry(original).CurrentValues.SetValues(novo);
                     context.SaveChanges();
                 }
+                //} catch { MessageBox.Show("Erro ao modificar produto"); }
                 Update(sender, null);
             }
         }
@@ -101,6 +130,20 @@ namespace View.UserControls {
             } else if (e.RemovedItems.Count > 0) {
                 Update(sender, null);
             }
+        }
+
+        private void ButtonEditar_Click(object sender, RoutedEventArgs e) {
+            if (DataGridProduto.SelectedItems.Count == 1) {
+                if (p == null) {
+                    using (var ctx = new ERPDBModelContainer()) {
+                        p = new AdicionarProdutoWindow(ctx.ProdutoSet.Single(o=>o.Id==((ProdutoAdapter)DataGridProduto.SelectedItem).Id));
+                    }
+                    p.Closed += (a, b) => p = null;
+                    p.buttonAdd.Click += Update;
+                    p.Show();
+                } else if (p.IsVisible) p.Focus();
+                else p.Show();
+            } else { MessageBox.Show("Você precisa selecionar apenas 1 Item"); }
         }
     }
 }
